fix(tests): cast config values to str and split oversized test file
